requirements.txt
black
flake8
isort
pytest
pre-commit
pytest-mock
mysql-connector-python
